pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.65",
    "fastapi>=0.135.1",
    "pika>=1.3.2",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
]
